+++ src/connected/AppBar.tsx
@@ -5,10 +5,11 @@ import { Theme, WithStyles, createStyles, withStyles } from '@material-ui/core';
 import SettingsApplicationsIcon from '@material-ui/icons/SettingsApplications';
 import deepOrange from '@material-ui/core/colors/deepOrange';
 import { AppBar, Toolbar, Typography, IconButton } from '@material-ui/core';
-// import { AppBar, Toolbar, Typography, IconButton, Badge } from '@material-ui/core';
-// import NotificationsIcon from '@material-ui/icons/Notifications';
-// import Avatar from '@material-ui/core/Avatar';
+import FaceIcon from '@material-ui/icons/Face';
+import Avatar from '@material-ui/core/Avatar';
+import CircularProgress from '@material-ui/core/CircularProgress';
 
+import { useAuth0 } from '../components/Auth0Provider';
 import { AppState } from '../states/appState';
 import logo from '../resources/logo.png';
 import { VIEWER_LOGO_WIDTH, VIEWER_HEADER_BACKGROUND_COLOR, VIEWER_APP_NAME } from '../config';
@@ -63,45 +64,95 @@ const styles = (theme: Theme) => createStyles(
             marginLeft: theme.spacing(1),
         },
         orangeAvatar: {
-            margin: 10,
             color: '#fff',
             backgroundColor: deepOrange[300],
         },
-    });
+        signInWrapper: {
+            margin: theme.spacing(1),
+            position: 'relative',
+        },
+        signInProgress: {
+            color: deepOrange[300],
+            position: 'absolute',
+            top: '50%',
+            left: '50%',
+            zIndex: 1,
+            marginTop: -12,
+            marginLeft: -12,
+        },
+    }
+);
 
-class _AppBar extends React.Component<AppBarProps> {
+const _AppBar: React.FC<AppBarProps> = ({classes, appName, openDialog}: AppBarProps) => {
 
-    handleSettingsButtonClicked = (event: React.MouseEvent<HTMLButtonElement>) => {
-        this.props.openDialog('settings');
+    const {isAuthenticated, user, loginWithRedirect, loading} = useAuth0();
+
+    const handleSettingsButtonClicked = (event: React.MouseEvent<HTMLButtonElement>) => {
+        openDialog('settings');
     };
 
-    render() {
-        const {classes, appName} = this.props;
-        return (
-            <AppBar
-                position="absolute"
-                className={classNames(classes.appBar)}
-            >
-                <Toolbar disableGutters={true} className={classes.toolbar}>
-                    <img src={logo} alt={'xcube logo'} width={VIEWER_LOGO_WIDTH} className={classes.logo}/>
-                    <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
-                        {appName}
-                    </Typography>
-                    {/*<IconButton>*/}
-                    {/*<Badge badgeContent={4} color={"secondary"}>*/}
-                    {/*<NotificationsIcon/>*/}
-                    {/*</Badge>*/}
-                    {/*</IconButton>*/}
-                    {/*<Avatar className={classes.orangeAvatar}>CL</Avatar>*/}
-                    <IconButton onClick={this.handleSettingsButtonClicked}>
-                        <SettingsApplicationsIcon/>
-                    </IconButton>
-                </Toolbar>
-                <ServerDialog/>
-                <SettingsDialog/>
-            </AppBar>
+    const handleSignInButtonCLicked = () => {
+        loginWithRedirect({});
+    };
+
+    let userButton;
+    if (isAuthenticated) {
+        let avatarContent: React.ReactNode = <FaceIcon/>;
+        if (!user) {
+            userButton = <Avatar className={classes.orangeAvatar}>?</Avatar>;
+        } else if (user.picture) {
+            userButton = <Avatar src={user.picture} alt={user.name}/>;
+        } else {
+            const name1 = user.given_name || user.name || user.nickname;
+            const name2 = user.family_name;
+            let letters: string | null = null;
+            if (name1 && name2) {
+                letters = name1[0] + name2[0];
+            } else if (name1) {
+                letters = name1[0];
+            } else if (name2) {
+                letters = name2[0];
+            }
+            if (letters !== null) {
+                avatarContent = letters.toUpperCase();
+            }
+            userButton = <Avatar className={classes.orangeAvatar}>{avatarContent}</Avatar>;
+        }
+    } else {
+        userButton = (
+            <IconButton onClick={loading ? undefined : handleSignInButtonCLicked}>
+                <FaceIcon/>
+            </IconButton>
         );
+        if (loading) {
+            userButton = (
+                <div className={classes.signInWrapper}>
+                    {userButton}
+                    <CircularProgress size={30} className={classes.signInProgress}/>
+                </div>
+            );
+        }
     }
-}
+
+    return (
+        <AppBar
+            position="absolute"
+            className={classNames(classes.appBar)}
+        >
+            <Toolbar disableGutters={true} className={classes.toolbar}>
+                <img src={logo} alt={'xcube logo'} width={VIEWER_LOGO_WIDTH} className={classes.logo}/>
+                <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
+                    {appName}
+                </Typography>
+                {userButton}
+                <IconButton onClick={handleSettingsButtonClicked}>
+                    <SettingsApplicationsIcon/>
+                </IconButton>
+            </Toolbar>
+            <ServerDialog/>
+            <SettingsDialog/>
+        </AppBar>
+    );
+};
 
 export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(_AppBar));